Initialize models just once, and reuse objs
Before this commit, new model objects were being created for each test.
Totally unnecessary.
